// web/src/client/models/ToolOutIdWithAndName.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ToolOutIdWithAndName = {
    id: number;
    name: string;
    description: string;
    display_name: (string | null);
    input_parameters: (Record<string, any> | null);
    is_online?: (boolean | null);
};